requirements.txt
# I'm building an environment using anaconda, but this is the list
# AZ requires a requirements.txt file

azure-functions
azure-eventhub
pyyaml 
azure-digitaltwins-core
azure-identity
pyOpenSSL